+++ ChangeLog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 
+### Fixed
+- Offset of accessor into buffer was incorrect for some scalar accessors (#262)
+
 ## [4.4.1] - 2021-10-27
 ### Fixed
 - .NET 4.6 compiler issue (#261)